+++ .gitignore
@@ -44,3 +44,5 @@ out
 /file-uploads
 /reader-assets
 /.lh/
+
+/simple-web/
