+++ web/src/plugins/element.js
@@ -30,7 +30,9 @@ import {
   Pagination,
   InputNumber,
   Image,
-  Badge
+  Badge,
+  Tabs,
+  TabPane
 } from "element-ui";
 
 Vue.use(Button);
@@ -62,6 +64,8 @@ Vue.use(Pagination);
 Vue.use(InputNumber);
 Vue.use(Image);
 Vue.use(Badge);
+Vue.use(Tabs);
+Vue.use(TabPane);
 
 Vue.prototype.$msgbox = MessageBox;
 Vue.prototype.$message = Object.assign({}, Message, {
